Fix for #14: Put a method into superclass, deleted javadoc, optimization
